chore: rename finder to pages finder
